adding seirs and seirs_seasonal_forcing examples as well as separating some tests
