+++ examples/sir_infer_parameters.py
@@ -12,30 +12,25 @@ from sir_age_stratified import SIRConfig, run_simulation
 from dynode import MCMCProcess, Strain, SVIProcess
 from dynode.config import SimulationConfig
 
-SHOW_PLOTS = False
 
-
-# define the entire process of simulating incidence
 def model(
     config: SimulationConfig,
     tf,
     obs_data: jax.Array | None = None,
-    infer_mode=False,
 ):
     """Numpyro model for simulating infection incidence of an SIR model."""
     solution: Solution = run_simulation(config, tf)
     # compare to observed data if we have it
-    if infer_mode:
-        assert solution.ys is not None, "mypy assert"
-        incidence = jnp.diff(
-            solution.ys[config.idx.r], axis=0
-        )  # leading time axis
-        incidence = jnp.maximum(incidence, 1e-6)
-        numpyro.sample(
-            "inf_incidence",
-            numpyro.distributions.Poisson(incidence),
-            obs=obs_data,
-        )
+    assert solution.ys is not None, "mypy assert"
+    incidence = jnp.diff(
+        solution.ys[config.idx.r], axis=0
+    )  # leading time axis
+    incidence = jnp.maximum(incidence, 1e-6)
+    numpyro.sample(
+        "inf_incidence",
+        numpyro.distributions.Poisson(incidence),
+        obs=obs_data,
+    )
     return solution
 
 
@@ -73,21 +68,20 @@ if __name__ == "__main__":
     assert solution.ys is not None
     idx = config_static.idx
     # add 1 to each axis to account for the leading time dimension in `solution`
-    if SHOW_PLOTS:
-        plt.plot(
-            jnp.sum(solution.ys[idx.s], axis=idx.s.age + 1),
-            label="s",
-        )
-        plt.plot(
-            jnp.sum(solution.ys[idx.i], axis=idx.i.age + 1),
-            label="i",
-        )
-        plt.plot(
-            jnp.sum(solution.ys[idx.r], axis=idx.r.age + 1),
-            label="r",
-        )
-        plt.legend()
-        plt.show()
+    plt.plot(
+        jnp.sum(solution.ys[idx.s], axis=idx.s.age + 1),
+        label="s",
+    )
+    plt.plot(
+        jnp.sum(solution.ys[idx.i], axis=idx.i.age + 1),
+        label="i",
+    )
+    plt.plot(
+        jnp.sum(solution.ys[idx.r], axis=idx.r.age + 1),
+        label="r",
+    )
+    plt.legend()
+    plt.show()
 
     # diff recovered individuals to recover lagged incidence for each age group
     incidence = jnp.diff(solution.ys[idx.r], axis=0)
@@ -112,13 +106,13 @@ if __name__ == "__main__":
     # running inference
     print("fitting MCMC")
     inferer_mcmc = inference_process_mcmc.infer(
-        config=config_infer, tf=100, obs_data=incidence, infer_mode=True
+        config=config_infer, tf=100, obs_data=incidence
     )
     posterior_samples_mcmc = inference_process_mcmc.get_samples()
     # %%
     print("fitting SVI")
     inferer_svi = inference_process_svi.infer(
-        config=config_infer, tf=100, obs_data=incidence, infer_mode=True
+        config=config_infer, tf=100, obs_data=incidence
     )
     posterior_samples_svi = inference_process_svi.get_samples()
 
@@ -138,30 +132,28 @@ if __name__ == "__main__":
         f"SVI posterior's R0: {jnp.mean(posterior_samples_svi['strains_0_r0'])}, "
         f"Infectious Period: {jnp.mean(posterior_samples_svi['strains_0_infectious_period'])}"
     )
-    # %%
-    if SHOW_PLOTS:
-        svi_arviz = inference_process_svi.to_arviz()
-        print(
-            "the following arviz object is only interactive if run as a notebook."
-        )
-        print(svi_arviz)
+    svi_arviz = inference_process_svi.to_arviz()
+    print(
+        "the following arviz object is only interactive if run as a notebook."
+    )
+    print(svi_arviz)
 
-        mcmc_arviz = inference_process_mcmc.to_arviz()
-        axes = az.plot_density(
-            [mcmc_arviz],
-            data_labels=["R0"],
-            var_names=["strains_0_r0"],
-            shade=0.2,
-        )
+    mcmc_arviz = inference_process_mcmc.to_arviz()
+    axes = az.plot_density(
+        [mcmc_arviz],
+        data_labels=["R0"],
+        var_names=["strains_0_r0"],
+        shade=0.2,
+    )
 
-        fig = axes.flatten()[0].get_figure()
-        fig.suptitle("Density Interval for R0")
+    fig = axes.flatten()[0].get_figure()
+    fig.suptitle("Density Interval for R0")
 
-        plt.show()
-        print(
-            "the following arviz object is only interactive if run as a notebook."
-        )
-        print(mcmc_arviz)
+    plt.show()
+    print(
+        "the following arviz object is only interactive if run as a notebook."
+    )
+    print(mcmc_arviz)
     # %%
     # projecting forward
     # now lets turn on Predictive mode and do some projections forward without observed data
@@ -175,7 +167,6 @@ if __name__ == "__main__":
         config=config_infer,  # arguments passed to `model`
         tf=200,
         obs_data=None,
-        infer_mode=True,
     )
     assert inference_process_svi._inferer is not None, "mypy assert"
     assert isinstance(inference_process_svi._inference_state, SVIRunResult)
@@ -190,38 +181,31 @@ if __name__ == "__main__":
         config=config_infer,  # arguments passed to `model`
         tf=200,
         obs_data=None,
-        infer_mode=True,
     )
 
-    # %%
-    if SHOW_PLOTS:
-        # pick a random subset of 50 samples and plot the incidence, plot the true incidence from earlier as well
-        random_samples = jax.random.choice(
-            inference_process_mcmc.inference_prngkey,
-            posterior_incidence_mcmc["inf_incidence"].shape[0],
-            shape=(50,),
+    # pick a random subset of 50 samples and plot the incidence, plot the true incidence from earlier as well
+    random_samples = jax.random.choice(
+        inference_process_mcmc.inference_prngkey,
+        posterior_incidence_mcmc["inf_incidence"].shape[0],
+        shape=(50,),
+    )
+    for sample in random_samples:
+        plt.plot(
+            jnp.sum(posterior_incidence_mcmc["inf_incidence"][sample], axis=1),
+            label=None,
         )
-        for sample in random_samples:
-            plt.plot(
-                jnp.sum(
-                    posterior_incidence_mcmc["inf_incidence"][sample], axis=1
-                ),
-                label=None,
-            )
-        plt.plot(jnp.sum(incidence, axis=1), label="true incidence")
-        plt.legend()
-        plt.title("MCMC posterior predictive")
-        plt.show()
+    plt.plot(jnp.sum(incidence, axis=1), label="true incidence")
+    plt.legend()
+    plt.title("MCMC posterior predictive")
+    plt.show()
 
-        for sample in random_samples:
-            plt.plot(
-                jnp.sum(
-                    posterior_incidence_svi["inf_incidence"][sample], axis=1
-                ),
-                label=None,
-            )
-        plt.plot(jnp.sum(incidence, axis=1), label="true incidence")
-        plt.legend()
-        plt.title("SVI posterior predictive")
-        plt.show()
+    for sample in random_samples:
+        plt.plot(
+            jnp.sum(posterior_incidence_svi["inf_incidence"][sample], axis=1),
+            label=None,
+        )
+    plt.plot(jnp.sum(incidence, axis=1), label="true incidence")
+    plt.legend()
+    plt.title("SVI posterior predictive")
+    plt.show()
     # %%
